feat: add nanogpt stream and replay functionality with anomaly logging
- Implemented replay warnings in `provider/replay/warnings.ts` for various anomaly scenarios.
- Created a bridge for handling tool calls and messages in `provider/stream/bridge.ts`.
- Added inspection utilities in `provider/stream/inspection.ts` to analyze stream content.
- Developed payload management functions in `provider/stream/payload.ts` for streaming responses.
- Established replay stream handling in `provider/stream/replay.ts` to manage assistant messages.
- Defined types for stream events and messages in `provider/stream/types.ts`.
- Introduced comprehensive warning mechanisms in `provider/stream/warnings.ts` to log anomalies during stream processing.
