+++ package.json
@@ -41,9 +41,20 @@
     "provider/inspection.ts",
     "provider/markers.ts",
     "provider/replay-hooks.ts",
+    "provider/replay/types.ts",
+    "provider/replay/inspection.ts",
+    "provider/replay/warnings.ts",
+    "provider/replay/sanitize.ts",
+    "provider/replay/validate.ts",
     "provider/tool-schema-hooks.ts",
     "provider/error-hooks.ts",
     "provider/stream-hooks.ts",
+    "provider/stream/types.ts",
+    "provider/stream/inspection.ts",
+    "provider/stream/payload.ts",
+    "provider/stream/bridge.ts",
+    "provider/stream/replay.ts",
+    "provider/stream/warnings.ts",
     "provider/bridge/types.ts",
     "provider/bridge/system-prompt.ts",
     "provider/bridge/object-parser.ts",
