+++ provider/stream-hooks.ts
@@ -1,915 +1,32 @@
-import { randomUUID } from "node:crypto";
-import type { NanoGptPluginConfig, NanoGptResponseFormat } from "../models.js";
+import type { NanoGptPluginConfig } from "../models.js";
 import type { AnyAgentTool, ProviderWrapStreamFnContext } from "openclaw/plugin-sdk/plugin-entry";
 import {
-  createNanoGptAnomalyWarnOnceLogger,
-  type NanoGptAnomalyWarning,
-} from "./anomaly-logger.js";
-import {
-  buildNanoGptExpectedShapeSummary,
-  buildNanoGptObservedShapeSummary,
   resolveNanoGptModelIdentity,
-  type NanoGptModelFamily,
 } from "./anomaly-types.js";
 import { isRecord } from "../shared/guards.js";
-import {
-  collectNanoGptStreamMarkerInspection,
-  type NanoGptStreamMarkerInspection,
-} from "./inspection.js";
-import { parseObjectBridgeAssistantText } from "./bridge/object-parser.js";
 import { buildNanoGptBridgeRetrySystemMessage } from "./bridge/retry.js";
+import { createNanoGptLoggerSync } from "./nanogpt-logger.js";
 import {
-  buildNanoGptObjectBridgeSystemMessage,
-  buildNanoGptXmlBridgeSystemMessage,
-} from "./bridge/system-prompt.js";
-import { parseXmlBridgeAssistantText } from "./bridge/xml-parser.js";
-import { createNanoGptLoggerSync, type NanoGptLogger } from "./nanogpt-logger.js";
-
-type NanoGptWrappedStreamFn = ProviderWrapStreamFnContext["streamFn"];
-type NanoGptStreamResult = Awaited<ReturnType<NonNullable<NanoGptWrappedStreamFn>>>;
-
-type NanoGptToolCall = Readonly<{
-  type: "toolCall";
-  id: string;
-  name: string;
-  arguments: Record<string, unknown>;
-}>;
-
-type NanoGptAssistantContentBlock =
-  | Readonly<{ type: "text"; text: string }>
-  | Readonly<{ type: "thinking"; thinking: string }>
-  | NanoGptToolCall
-  | (Record<string, unknown> & { type: string });
-
-type NanoGptAssistantMessage = Readonly<{
-  role: string;
-  content: NanoGptAssistantContentBlock[];
-  stopReason: string;
-  usage?: NanoGptUsage;
-  api?: string;
-  provider?: string;
-  model?: string;
-  timestamp?: number;
-}>;
-
-type NanoGptReplayEvent =
-  | Readonly<{ type: "start"; partial: NanoGptAssistantMessage }>
-  | Readonly<{ type: "text_start"; contentIndex: number; partial: NanoGptAssistantMessage }>
-  | Readonly<{ type: "text_delta"; contentIndex: number; delta: string; partial: NanoGptAssistantMessage }>
-  | Readonly<{ type: "text_end"; contentIndex: number; content: string; partial: NanoGptAssistantMessage }>
-  | Readonly<{ type: "thinking_start"; contentIndex: number; partial: NanoGptAssistantMessage }>
-  | Readonly<{ type: "thinking_delta"; contentIndex: number; delta: string; partial: NanoGptAssistantMessage }>
-  | Readonly<{ type: "thinking_end"; contentIndex: number; content: string; partial: NanoGptAssistantMessage }>
-  | Readonly<{ type: "toolcall_start"; contentIndex: number; partial: NanoGptAssistantMessage }>
-  | Readonly<{ type: "toolcall_delta"; contentIndex: number; delta: string; partial: NanoGptAssistantMessage }>
-  | Readonly<{
-      type: "toolcall_end";
-      contentIndex: number;
-      toolCall: NanoGptToolCall;
-      partial: NanoGptAssistantMessage;
-    }>
-  | Readonly<{
-      type: "done";
-      reason: "stop" | "length" | "toolUse";
-      message: NanoGptAssistantMessage;
-    }>;
-
-type NanoGptReplayStream = {
-  push: (event: NanoGptReplayEvent) => void;
-  end: (message?: NanoGptAssistantMessage) => void;
-  result: () => Promise<NanoGptAssistantMessage>;
-  [Symbol.asyncIterator]: () => AsyncIterator<NanoGptReplayEvent>;
-};
-
-type NanoGptPluginLogger = {
-  warn?: (message: string, meta?: Record<string, unknown>) => void;
-};
-
-type NanoGptRequestToolMetadata = Readonly<{
-  toolEnabled: boolean;
-  toolCount: number;
-  toolNames: readonly string[];
-}>;
-
-type NanoGptStreamContentInspection = Readonly<{
-  visibleText: string;
-  visibleTextLength: number;
-  textBlockCount: number;
-  toolCallCount: number;
-  thinkingBlockCount: number;
-}>;
-
-type NanoGptUsage = {
-  input: number;
-  output: number;
-  cacheRead: number;
-  cacheWrite: number;
-  totalTokens: number;
-  cost: {
-    input: number;
-    output: number;
-    cacheRead: number;
-    cacheWrite: number;
-    total: number;
-  };
-};
-
-const NANO_GPT_STREAM_ANOMALY_LOGGER_CACHE = new WeakMap<
+  buildNanoGptBridgeFailureMessage,
+  rewriteNanoGptBridgeMessage,
+} from "./stream/bridge.js";
+import { collectNanoGptRequestToolMetadata } from "./stream/inspection.js";
+import {
+  ensureIncludeUsageInStreamingPayload,
+  injectNanoGptBridgePayload,
+  maybeInjectNanoGptResponseFormat,
+  resolveNanoGptBridgeProtocol,
+  shouldApplyNanoGptBridge,
+} from "./stream/payload.js";
+import { replayNanoGptAssistantMessage } from "./stream/replay.js";
+import type {
   NanoGptPluginLogger,
-  (warning: NanoGptAnomalyWarning) => void
->();
-
-function collectNanoGptRequestToolMetadata(context: unknown): NanoGptRequestToolMetadata {
-  if (!isRecord(context) || !Array.isArray(context.tools)) {
-    return {
-      toolEnabled: false,
-      toolCount: 0,
-      toolNames: [],
-    };
-  }
-
-  const toolNames = new Set<string>();
-  for (const tool of context.tools) {
-    if (!isRecord(tool) || typeof tool.name !== "string") {
-      continue;
-    }
-
-    const normalizedName = tool.name.trim();
-    if (normalizedName) {
-      toolNames.add(normalizedName);
-    }
-  }
-
-  return {
-    toolEnabled: context.tools.length > 0,
-    toolCount: context.tools.length,
-    toolNames: [...toolNames],
-  };
-}
-
-function collectNanoGptStreamContentInspection(finalMessage: unknown): NanoGptStreamContentInspection | null {
-  if (!isRecord(finalMessage) || !Array.isArray(finalMessage.content)) {
-    return null;
-  }
-
-  let visibleText = "";
-  let textBlockCount = 0;
-  let toolCallCount = 0;
-  let thinkingBlockCount = 0;
-
-  for (const contentBlock of finalMessage.content) {
-    if (!isRecord(contentBlock) || typeof contentBlock.type !== "string") {
-      continue;
-    }
-
-    if (contentBlock.type === "text") {
-      textBlockCount += 1;
-      if (typeof contentBlock.text === "string") {
-        visibleText += contentBlock.text;
-      }
-      continue;
-    }
-
-    if (contentBlock.type === "toolCall") {
-      toolCallCount += 1;
-      continue;
-    }
-
-    if (contentBlock.type === "thinking") {
-      thinkingBlockCount += 1;
-    }
-  }
-
-  return {
-    visibleText,
-    visibleTextLength: visibleText.trim().length,
-    textBlockCount,
-    toolCallCount,
-    thinkingBlockCount,
-  };
-}
-
-function createNanoGptStreamAnomalyLogger(logger?: NanoGptPluginLogger) {
-  if (!logger?.warn) {
-    return undefined;
-  }
-
-  const cachedLogger = NANO_GPT_STREAM_ANOMALY_LOGGER_CACHE.get(logger);
-  if (cachedLogger) {
-    return cachedLogger;
-  }
-
-  const warnOnceLogger = createNanoGptAnomalyWarnOnceLogger({
-    logger: {
-      warn: (message: string) => logger.warn?.(message),
-    },
-  });
-
-  NANO_GPT_STREAM_ANOMALY_LOGGER_CACHE.set(logger, warnOnceLogger);
-  return warnOnceLogger;
-}
-
-function buildNanoGptExpectedToolRequestShapeSummary(requestToolMetadata: NanoGptRequestToolMetadata) {
-  return buildNanoGptExpectedShapeSummary({
-    headline: requestToolMetadata.toolEnabled ? "tool-enabled request" : "tool-free request",
-    counts: {
-      toolCount: requestToolMetadata.toolCount,
-    },
-    groups:
-      requestToolMetadata.toolNames.length > 0
-        ? [{ label: "toolNames", values: requestToolMetadata.toolNames }]
-        : undefined,
-  });
-}
-
-function buildNanoGptObservedStreamShapeSummary(params: {
-  headline: string;
-  counts?: Record<string, number>;
-  markerNames?: readonly string[];
-}) {
-  return buildNanoGptObservedShapeSummary({
-    headline: params.headline,
-    counts: params.counts,
-    groups:
-      params.markerNames && params.markerNames.length > 0
-        ? [{ label: "markers", values: params.markerNames }]
-        : undefined,
-  });
-}
-
-function emitNanoGptStreamAnomaly(params: {
-  warnNanoGptAnomaly: (warning: NanoGptAnomalyWarning) => void;
-  kind: NanoGptAnomalyWarning["kind"];
-  modelId: string;
-  modelFamily: NanoGptModelFamily;
-  transportApi?: string;
-  requestToolMetadata: NanoGptRequestToolMetadata;
-  observedHeadline: string;
-  observedCounts?: Record<string, number>;
-  observedMarkerNames?: readonly string[];
-  finishReason?: string;
-}): void {
-  params.warnNanoGptAnomaly({
-    kind: params.kind,
-    stage: "stream_result",
-    providerId: "nanogpt",
-    modelId: params.modelId,
-    modelFamily: params.modelFamily,
-    transportApi: params.transportApi,
-    expectedShapeSummary: buildNanoGptExpectedToolRequestShapeSummary(params.requestToolMetadata),
-    observedShapeSummary: buildNanoGptObservedStreamShapeSummary({
-      headline: params.observedHeadline,
-      counts: params.observedCounts,
-      markerNames: params.observedMarkerNames,
-    }),
-    metadata: {
-      toolNames: params.requestToolMetadata.toolNames,
-      toolCount: params.requestToolMetadata.toolCount,
-      finishReason: params.finishReason,
-      markerNames: params.observedMarkerNames,
-    },
-  });
-}
-
-function shouldWarnNanoGptToolEnabledTurnWithoutToolCall(params: {
-  inspection: NanoGptStreamContentInspection;
-  markerInspection: NanoGptStreamMarkerInspection;
-}): boolean {
-  return params.inspection.visibleTextLength === 0 || params.markerInspection.toolLikeMarkers.length > 0;
-}
-
-function scheduleNanoGptStreamResultWarnings(params: {
-  stream: unknown;
-  logger?: NanoGptPluginLogger;
-  nanogptLogger?: NanoGptLogger;
-  warnNanoGptAnomaly?: (warning: NanoGptAnomalyWarning) => void;
-  modelId: string;
-  modelFamily: NanoGptModelFamily;
-  transportApi?: string;
-  requestedIncludeUsage: boolean;
-  requestToolMetadata: NanoGptRequestToolMetadata;
-}): void {
-  if (!params.stream || typeof (params.stream as any).result !== "function") {
-    return;
-  }
-
-  void (params.stream as any)
-    .result()
-    .then((finalMessage: unknown) => {
-      if (params.requestedIncludeUsage && isRecord(finalMessage)) {
-        const { empty, invalidFields } = inspectUsage(finalMessage.usage);
-        if (empty || invalidFields.length > 0) {
-          params.logger?.warn?.(
-            `[nanogpt] requested stream_options.include_usage but received ${empty ? "empty" : "invalid"} usage in stream result`,
-            {
-              modelId: params.modelId,
-              ...(invalidFields.length > 0 ? { invalidFields } : {}),
-            },
-          );
-          params.nanogptLogger?.warn(
-            `[nanogpt] requested stream_options.include_usage but received ${empty ? "empty" : "invalid"} usage in stream result`,
-            {
-              modelId: params.modelId,
-              ...(invalidFields.length > 0 ? { invalidFields } : {}),
-            },
-          );
-        }
-      }
-
-      if (!params.warnNanoGptAnomaly) {
-        return;
-      }
-
-      const inspection = collectNanoGptStreamContentInspection(finalMessage);
-      if (!inspection) {
-        return;
-      }
-
-      const markerInspection = collectNanoGptStreamMarkerInspection(inspection.visibleText);
-      const finishReason = isRecord(finalMessage) && typeof finalMessage.stopReason === "string"
-        ? finalMessage.stopReason
-        : undefined;
-
-      if (
-        params.requestToolMetadata.toolEnabled &&
-        inspection.toolCallCount === 0 &&
-        shouldWarnNanoGptToolEnabledTurnWithoutToolCall({
-          inspection,
-          markerInspection,
-        })
-      ) {
-        emitNanoGptStreamAnomaly({
-          warnNanoGptAnomaly: params.warnNanoGptAnomaly,
-          kind: "tool_enabled_turn_without_tool_call",
-          modelId: params.modelId,
-          modelFamily: params.modelFamily,
-          transportApi: params.transportApi,
-          requestToolMetadata: params.requestToolMetadata,
-          observedHeadline: "assistant result without parsed tool calls",
-          observedCounts: {
-            parsedToolCalls: inspection.toolCallCount,
-            textBlocks: inspection.textBlockCount,
-            thinkingBlocks: inspection.thinkingBlockCount,
-            visibleTextLength: inspection.visibleTextLength,
-          },
-          finishReason,
-        });
-
-        if (inspection.visibleTextLength === 0) {
-          emitNanoGptStreamAnomaly({
-            warnNanoGptAnomaly: params.warnNanoGptAnomaly,
-            kind: "tool_enabled_turn_with_empty_visible_output",
-            modelId: params.modelId,
-            modelFamily: params.modelFamily,
-            transportApi: params.transportApi,
-            requestToolMetadata: params.requestToolMetadata,
-            observedHeadline: "tool-enabled turn with empty visible assistant output",
-            observedCounts: {
-              parsedToolCalls: inspection.toolCallCount,
-              textBlocks: inspection.textBlockCount,
-              thinkingBlocks: inspection.thinkingBlockCount,
-              visibleTextLength: inspection.visibleTextLength,
-            },
-            finishReason,
-          });
-        }
-
-        if (markerInspection.toolLikeMarkers.length > 0) {
-          emitNanoGptStreamAnomaly({
-            warnNanoGptAnomaly: params.warnNanoGptAnomaly,
-            kind: "tool_enabled_turn_with_tool_like_text",
-            modelId: params.modelId,
-            modelFamily: params.modelFamily,
-            transportApi: params.transportApi,
-            requestToolMetadata: params.requestToolMetadata,
-            observedHeadline: "tool-enabled turn with tool-like visible text",
-            observedCounts: {
-              parsedToolCalls: inspection.toolCallCount,
-              textBlocks: inspection.textBlockCount,
-              thinkingBlocks: inspection.thinkingBlockCount,
-              visibleTextLength: inspection.visibleTextLength,
-              toolLikeMarkerCount: markerInspection.toolLikeMarkers.length,
-            },
-            observedMarkerNames: markerInspection.toolLikeMarkers,
-            finishReason,
-          });
-        }
-      }
-
-      if (markerInspection.reasoningIsUnbalanced) {
-        emitNanoGptStreamAnomaly({
-          warnNanoGptAnomaly: params.warnNanoGptAnomaly,
-          kind: "visible_output_contains_unbalanced_reasoning_tags",
-          modelId: params.modelId,
-          modelFamily: params.modelFamily,
-          transportApi: params.transportApi,
-          requestToolMetadata: params.requestToolMetadata,
-          observedHeadline: "visible assistant text contains unbalanced reasoning tags",
-          observedCounts: {
-            parsedToolCalls: inspection.toolCallCount,
-            textBlocks: inspection.textBlockCount,
-            thinkingBlocks: inspection.thinkingBlockCount,
-            visibleTextLength: inspection.visibleTextLength,
-            reasoningMarkerCount: markerInspection.reasoningMarkerNames.length,
-          },
-          observedMarkerNames: markerInspection.reasoningMarkerNames,
-          finishReason,
-        });
-      } else if (markerInspection.reasoningMarkerNames.length > 0) {
-        emitNanoGptStreamAnomaly({
-          warnNanoGptAnomaly: params.warnNanoGptAnomaly,
-          kind: "visible_output_contains_reasoning_tags",
-          modelId: params.modelId,
-          modelFamily: params.modelFamily,
-          transportApi: params.transportApi,
-          requestToolMetadata: params.requestToolMetadata,
-          observedHeadline: "visible assistant text contains reasoning tags",
-          observedCounts: {
-            parsedToolCalls: inspection.toolCallCount,
-            textBlocks: inspection.textBlockCount,
-            thinkingBlocks: inspection.thinkingBlockCount,
-            visibleTextLength: inspection.visibleTextLength,
-            reasoningMarkerCount: markerInspection.reasoningMarkerNames.length,
-          },
-          observedMarkerNames: markerInspection.reasoningMarkerNames,
-          finishReason,
-        });
-      }
-
-      if (markerInspection.xmlLikeToolWrapperMarkers.length > 0) {
-        emitNanoGptStreamAnomaly({
-          warnNanoGptAnomaly: params.warnNanoGptAnomaly,
-          kind: "visible_output_contains_xml_like_tool_wrappers",
-          modelId: params.modelId,
-          modelFamily: params.modelFamily,
-          transportApi: params.transportApi,
-          requestToolMetadata: params.requestToolMetadata,
-          observedHeadline: "visible assistant text contains XML-like tool wrappers",
-          observedCounts: {
-            parsedToolCalls: inspection.toolCallCount,
-            textBlocks: inspection.textBlockCount,
-            thinkingBlocks: inspection.thinkingBlockCount,
-            visibleTextLength: inspection.visibleTextLength,
-            xmlLikeToolWrapperCount: markerInspection.xmlLikeToolWrapperMarkers.length,
-          },
-          observedMarkerNames: markerInspection.xmlLikeToolWrapperMarkers,
-          finishReason,
-        });
-      }
-
-      if (markerInspection.functionCallMarkers.length > 0) {
-        emitNanoGptStreamAnomaly({
-          warnNanoGptAnomaly: params.warnNanoGptAnomaly,
-          kind: "visible_output_contains_function_call_markers",
-          modelId: params.modelId,
-          modelFamily: params.modelFamily,
-          transportApi: params.transportApi,
-          requestToolMetadata: params.requestToolMetadata,
-          observedHeadline: "visible assistant text contains function-call markers",
-          observedCounts: {
-            parsedToolCalls: inspection.toolCallCount,
-            textBlocks: inspection.textBlockCount,
-            thinkingBlocks: inspection.thinkingBlockCount,
-            visibleTextLength: inspection.visibleTextLength,
-            functionCallMarkerCount: markerInspection.functionCallMarkers.length,
-          },
-          observedMarkerNames: markerInspection.functionCallMarkers,
-          finishReason,
-        });
-      }
-    })
-    .catch(() => {
-      // Non-blocking: stream warnings are best-effort.
-    });
-}
-
-function isFiniteNonNegativeNumber(value: unknown): value is number {
-  return typeof value === "number" && Number.isFinite(value) && value >= 0;
-}
-
-function inspectUsage(usage: unknown): { empty: boolean; invalidFields: string[] } {
-  if (!isRecord(usage)) {
-    return { empty: true, invalidFields: ["usage"] };
-  }
-
-  const invalidFields: string[] = [];
-  const numericFields: Array<keyof NanoGptUsage> = [
-    "input",
-    "output",
-    "cacheRead",
-    "cacheWrite",
-    "totalTokens",
-  ];
-  for (const field of numericFields) {
-    if (!isFiniteNonNegativeNumber(usage[field])) {
-      invalidFields.push(`usage.${field}`);
-    }
-  }
-
-  if (!isRecord(usage.cost)) {
-    invalidFields.push("usage.cost");
-  } else {
-    const costFields: Array<keyof NanoGptUsage["cost"]> = [
-      "input",
-      "output",
-      "cacheRead",
-      "cacheWrite",
-      "total",
-    ];
-    for (const field of costFields) {
-      if (!isFiniteNonNegativeNumber(usage.cost[field])) {
-        invalidFields.push(`usage.cost.${field}`);
-      }
-    }
-  }
-
-  const empty =
-    invalidFields.length === 0 &&
-    usage.input === 0 &&
-    usage.output === 0 &&
-    usage.cacheRead === 0 &&
-    usage.cacheWrite === 0 &&
-    usage.totalTokens === 0 &&
-    isRecord(usage.cost) &&
-    usage.cost.input === 0 &&
-    usage.cost.output === 0 &&
-    usage.cost.cacheRead === 0 &&
-    usage.cost.cacheWrite === 0 &&
-    usage.cost.total === 0;
-
-  return { empty, invalidFields };
-}
-
-function ensureIncludeUsageInStreamingPayload(
-  payload: unknown,
-  forceIncludeUsage = true,
-): { payload?: unknown; requested: boolean } {
-  if (!isRecord(payload)) {
-    return { requested: false };
-  }
-
-  const streamValue = payload.stream;
-  const isStreaming = streamValue === true || streamValue === "true";
-  const hasStreamOptionsKey = "stream_options" in payload;
-  if (!isStreaming && !hasStreamOptionsKey) {
-    return { requested: false };
-  }
-
-  const existingStreamOptions = isRecord(payload.stream_options) ? payload.stream_options : undefined;
-  const existingIncludeUsage = existingStreamOptions?.include_usage;
-  if (existingIncludeUsage === true) {
-    return { requested: true };
-  }
-
-  if (!forceIncludeUsage) {
-    return {
-      requested: false,
-      payload,
-    };
-  }
-
-  return {
-    requested: true,
-    payload: {
-      ...payload,
-      stream_options: {
-        ...(existingStreamOptions ? existingStreamOptions : {}),
-        include_usage: true,
-      },
-    },
-  };
-}
-
-function hasParsedToolCalls(finalMessage: unknown): boolean {
-  return Boolean(collectNanoGptStreamContentInspection(finalMessage)?.toolCallCount);
-}
-
-function resolveNanoGptBridgeProtocol(
-  config: NanoGptPluginConfig | undefined,
-): "object" | "xml" {
-  return config?.bridgeProtocol === "xml" ? "xml" : "object";
-}
-
-function shouldApplyNanoGptBridge(
-  config: NanoGptPluginConfig | undefined,
-  requestToolMetadata: NanoGptRequestToolMetadata,
-): boolean {
-  return config?.bridgeMode === "always" && requestToolMetadata.toolEnabled;
-}
-
-function resolveNanoGptReplayStopReason(stopReason: NanoGptAssistantMessage["stopReason"]): "stop" | "length" | "toolUse" {
-  if (stopReason === "toolUse") {
-    return "toolUse";
-  }
-  if (stopReason === "length") {
-    return "length";
-  }
-  return "stop";
-}
-
-function resolveNanoGptBridgeStopReason(
-  parsedKind: "tool_calls" | "final",
-  stopReason: NanoGptAssistantMessage["stopReason"],
-): string {
-  if (parsedKind === "tool_calls") {
-    return "toolUse";
-  }
-  return stopReason === "toolUse" ? "stop" : stopReason;
-}
-
-function maybeInjectNanoGptResponseFormat(
-  payload: unknown,
-  responseFormat?: NanoGptResponseFormat,
-): unknown {
-  if (!responseFormat || !isRecord(payload)) {
-    return payload;
-  }
-
-  const existing = payload.response_format;
-  if (existing) {
-    return payload;
-  }
-
-  if (responseFormat === "json_object") {
-    return { ...payload, response_format: { type: "json_object" } };
-  }
-  if (typeof responseFormat === "object" && responseFormat.type === "json_schema") {
-    return {
-      ...payload,
-      response_format: responseFormat.schema
-        ? { type: "json_schema", json_schema: { schema: responseFormat.schema } }
-        : { type: "json_schema" },
-    };
-  }
-  return payload;
-}
-
-function injectNanoGptBridgePayload(params: {
-  payload: unknown;
-  tools: readonly AnyAgentTool[];
-  protocol: "object" | "xml";
-  retryMessage?: string;
-}): unknown {
-  if (!isRecord(params.payload)) {
-    return params.payload;
-  }
-
-  const messages = Array.isArray(params.payload.messages) ? [...params.payload.messages] : [];
-  const parallelAllowed = params.payload.parallel_tool_calls !== false;
-  const bridgeSystemMessage =
-    params.protocol === "xml"
-      ? buildNanoGptXmlBridgeSystemMessage(params.tools, parallelAllowed)
-      : buildNanoGptObjectBridgeSystemMessage(params.tools, parallelAllowed);
-
-  return {
-    ...params.payload,
-    messages: [
-      { role: "system", content: bridgeSystemMessage },
-      ...messages,
-      ...(params.retryMessage ? [{ role: "system", content: params.retryMessage }] : []),
-    ],
-  };
-}
-
-function buildNanoGptBridgeFailureMessage(finalMessage: NanoGptAssistantMessage): NanoGptAssistantMessage {
-  return {
-    ...finalMessage,
-    content: [
-      ...finalMessage.content.filter((block) => block.type === "thinking"),
-      {
-        type: "text",
-        text: "[nanogpt bridge] upstream returned no visible content or tool call for a tool-enabled turn.",
-      },
-    ],
-    stopReason: "stop",
-  };
-}
-
-function buildNanoGptBridgeToolCall(toolCall: { name: string; arguments: Record<string, unknown> }): NanoGptToolCall {
-  return {
-    type: "toolCall",
-    id: `call_${randomUUID().slice(0, 8)}`,
-    name: toolCall.name,
-    arguments: toolCall.arguments,
-  };
-}
-
-function rewriteNanoGptBridgeMessage(params: {
-  finalMessage: NanoGptAssistantMessage;
-  protocol: "object" | "xml";
-  tools: readonly AnyAgentTool[];
-}): NanoGptAssistantMessage | null {
-  if (hasParsedToolCalls(params.finalMessage)) {
-    return params.finalMessage;
-  }
-
-  const inspection = collectNanoGptStreamContentInspection(params.finalMessage);
-  const visibleText = inspection?.visibleText ?? "";
-  const parsed =
-    params.protocol === "xml"
-      ? parseXmlBridgeAssistantText(visibleText, params.tools)
-      : parseObjectBridgeAssistantText(visibleText, params.tools);
-
-  if (parsed.kind === "invalid") {
-    const trimmedVisibleText = visibleText.trim();
-    if (
-      trimmedVisibleText.length === 0 ||
-      parsed.error.code === "invalid_empty_turn" ||
-      parsed.error.code === "invalid_schema_turn" ||
-      (params.protocol === "object" && trimmedVisibleText.startsWith("{")) ||
-      (params.protocol === "xml" && trimmedVisibleText.includes("<"))
-    ) {
-      return null;
-    }
-    return params.finalMessage;
-  }
-
-  const content: NanoGptAssistantMessage["content"] = params.finalMessage.content.filter(
-    (block) => block.type === "thinking",
-  );
-  if (parsed.content) {
-    content.push({ type: "text", text: parsed.content });
-  }
-  if (parsed.kind === "tool_calls") {
-    content.push(...parsed.toolCalls.map(buildNanoGptBridgeToolCall));
-  }
-
-  return {
-    ...params.finalMessage,
-    content,
-    stopReason: resolveNanoGptBridgeStopReason(parsed.kind, params.finalMessage.stopReason),
-  };
-}
-
-function createNanoGptReplayStream(): NanoGptReplayStream {
-  const queuedEvents: NanoGptReplayEvent[] = [];
-  const iteratorWaiters: Array<(entry: IteratorResult<NanoGptReplayEvent>) => void> = [];
-  let streamClosed = false;
-  let finalMessage: NanoGptAssistantMessage | undefined;
-
-  let resolveResult!: (message: NanoGptAssistantMessage) => void;
-  let rejectResult!: (error: Error) => void;
-  let resultSettled = false;
-
-  const resultPromise = new Promise<NanoGptAssistantMessage>((resolve, reject) => {
-    resolveResult = resolve;
-    rejectResult = reject;
-  });
-
-  const settleResult = (message: NanoGptAssistantMessage) => {
-    if (resultSettled) {
-      return;
-    }
-    resultSettled = true;
-    resolveResult(message);
-  };
-
-  const stream: NanoGptReplayStream = {
-    push(event) {
-      if (streamClosed) {
-        return;
-      }
-
-      if (event.type === "done") {
-        finalMessage = event.message;
-        settleResult(event.message);
-      }
-
-      const waiter = iteratorWaiters.shift();
-      if (waiter) {
-        waiter({ value: event, done: false });
-        return;
-      }
-
-      queuedEvents.push(event);
-    },
-    end(message) {
-      if (streamClosed) {
-        return;
-      }
-
-      streamClosed = true;
-      if (message) {
-        finalMessage = message;
-        settleResult(message);
-      } else if (finalMessage) {
-        settleResult(finalMessage);
-      } else if (!resultSettled) {
-        resultSettled = true;
-        rejectResult(new Error("Replay stream ended without a final assistant message."));
-      }
-
-      while (iteratorWaiters.length > 0) {
-        const waiter = iteratorWaiters.shift();
-        waiter?.({ value: undefined, done: true });
-      }
-    },
-    result() {
-      return resultPromise;
-    },
-    [Symbol.asyncIterator]() {
-      return {
-        next: async () => {
-          if (queuedEvents.length > 0) {
-            const event = queuedEvents.shift();
-            if (!event) {
-              return { value: undefined, done: true };
-            }
-            return { value: event, done: false };
-          }
-
-          if (streamClosed) {
-            return { value: undefined, done: true };
-          }
-
-          return await new Promise<IteratorResult<NanoGptReplayEvent>>((resolve) => {
-            iteratorWaiters.push(resolve);
-          });
-        },
-      };
-    },
-  };
-
-  return stream;
-}
-
-function replayNanoGptAssistantMessage(message: NanoGptAssistantMessage): NanoGptStreamResult {
-  const stream = createNanoGptReplayStream();
-
-  queueMicrotask(() => {
-    stream.push({ type: "start", partial: message });
-    message.content.forEach((contentBlock, contentIndex) => {
-      if (contentBlock.type === "text" && typeof contentBlock.text === "string") {
-        stream.push({ type: "text_start", contentIndex, partial: message });
-        stream.push({
-          type: "text_delta",
-          contentIndex,
-          delta: contentBlock.text,
-          partial: message,
-        });
-        stream.push({
-          type: "text_end",
-          contentIndex,
-          content: contentBlock.text,
-          partial: message,
-        });
-        return;
-      }
-
-      if (contentBlock.type === "thinking" && typeof contentBlock.thinking === "string") {
-        stream.push({ type: "thinking_start", contentIndex, partial: message });
-        stream.push({
-          type: "thinking_delta",
-          contentIndex,
-          delta: contentBlock.thinking,
-          partial: message,
-        });
-        stream.push({
-          type: "thinking_end",
-          contentIndex,
-          content: contentBlock.thinking,
-          partial: message,
-        });
-        return;
-      }
-
-      if (
-        contentBlock.type === "toolCall" &&
-        typeof contentBlock.id === "string" &&
-        typeof contentBlock.name === "string" &&
-        isRecord(contentBlock.arguments)
-      ) {
-        const delta = JSON.stringify(contentBlock.arguments);
-        stream.push({ type: "toolcall_start", contentIndex, partial: message });
-        stream.push({
-          type: "toolcall_delta",
-          contentIndex,
-          delta,
-          partial: message,
-        });
-        stream.push({
-          type: "toolcall_end",
-          contentIndex,
-          toolCall: {
-            type: "toolCall",
-            id: contentBlock.id,
-            name: contentBlock.name,
-            arguments: contentBlock.arguments,
-          },
-          partial: message,
-        });
-      }
-    });
-    stream.push({
-      type: "done",
-      reason: resolveNanoGptReplayStopReason(message.stopReason),
-      message,
-    });
-    stream.end();
-  });
-
-  return stream as unknown as NanoGptStreamResult;
-}
+  NanoGptWrappedStreamFn,
+} from "./stream/types.js";
+import {
+  createNanoGptStreamAnomalyLogger,
+  scheduleNanoGptStreamResultWarnings,
+} from "./stream/warnings.js";
 
 export function wrapNanoGptStreamFn(
   ctx: ProviderWrapStreamFnContext,
